need user login and regestration message create update and all messages posted by soeone

+++ src/main/java/com/example/service/MessageService.java
@@ -6,6 +6,7 @@ import com.example.repository.MessageRepository;
 import com.example.entity.Message;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MessageService {
@@ -18,10 +19,14 @@ public class MessageService {
     }
 
     //TODO: create message
-    public Message createMessage(String Message_Text){
-        if ((Message_Text.isBlank() == true || Message_Text.length() > 255)) return null;
+    public Message createMessage(String messageText){
+        if ((messageText.isBlank() == true || messageText.length() > 255)) return null;
         else{
-            return null;
+            Message message = new Message();
+            message.setMessageText(messageText);
+            message.getPostedBy();
+            message.getTimePostedEpoch();
+            return message;
         }
     }
 
@@ -36,10 +41,27 @@ public class MessageService {
     }
 
     //TODO: delete message by message ID
-    public void deleteMessage(int messageID){
-        messageRepository.deleteById(messageID);
+    public int deleteMessage(int messageID){
+        Optional<Message> message = messageRepository.findById(messageID);
+
+        if(message.isPresent()){
+            messageRepository.deleteById(messageID);
+            return 1;
+        }
+        else return 0;
     }
 
-    //TODO: update message bu message ID
-    
+    //TODO: update message by message ID
+    public int updateMessage(int messageID, String messageText){
+        Optional<Message> messageOpt = messageRepository.findById(messageID);
+
+        if ((messageText.isBlank() == true || messageText.length() > 255 || !messageOpt.isPresent())) return 0;
+        else if(messageOpt.isPresent()){
+            Message existingMessage = messageOpt.get();
+            existingMessage.setMessageText(messageText);
+            messageRepository.save(existingMessage);
+            return 1;
+        }
+        else return 0;
+    }
 }
